Update ML_2016_SI_eusilc_cs.do
In your notes you wrote that I should update it to (105-15)/7 because of 15 days of compulsory prenatal leave. But the text goes as followed: "Use of 15 days before AND/OR AFTER is mandatory." So you can't say that all women HAVE to pick up 15 prenatal days off right? They could be postnatal too.
Also, I only changed ml_dur2 according to your notes, I didn't touch ml_dur1 because I was unsure if it needed to be updated to 15 if we assume the 15 days mandatory of prenatal leave.