Update PL_2017_NO_eusilc_cs.do
I double checked and it is actually that high: a ceiling of €9735/month.
I also double checked the flat-rate payment for non-employed women, and I can't believe the info in LP&R 2017. It says: "As of 1 January 2017, non-employed women receive a flat-rate payment of NOK61,120[€6,427] per child per month." --> €6427 per month sounds absurdly high, so i didn't code it... I think it must be an error of LP&R.
If you look at one year before it says: "Non-employed women receive a flat-rate payment of NOK46,000 [€4,951]." --> for the whole period I suppose.
If you look at one year after it says: "As of 1 January 2018, non-employed women receive a flat-rate payment of NOK63,140 [€6,696.95] per child." --> for the whole period I suppose.