Update PT_2012_SI_eusilc_cs.do
In LP & R 2012 (also later years) they say: "For the remaining 75 days the father is paid social security contributions based on the minimum wage (approximately €169 per month)". I wasn't sure if I needed to code this (I didn't code this in the previous years) but in the end I did it anyway since everyone still gets paid €169 instead of nothing. Tell me if this was not necessary. Also check the code, I'm not too sure about this. :)